requirements.txt
# Plugin dependencies
# Fixed versions based on current installation
aiohttp
jmcomic
img2pdf
ncatbot5
